Review: fixes for 1 review point

+++ src/api/v1/role/routes.py
@@ -2,9 +2,10 @@ from http import HTTPStatus
 
 from dependency_injector.wiring import Provide, inject
 from flask import Blueprint, Response, jsonify, make_response, request
+from flask_jwt_extended import jwt_required
 
 from core.containers import Container
-from core.utils import ServiceException
+from core.utils import ServiceException, authenticate
 from models.permission import Permission
 from services.role import RoleService
 
@@ -12,9 +13,14 @@ role = Blueprint('role', __name__, url_prefix='/role')
 
 
 @role.route('/', methods=['GET'])
+@jwt_required()
+@authenticate()
 @inject
-def get_roles(role_service: RoleService = Provide[Container.role_service]):
+def get_roles(
+        user_id: str,
+        role_service: RoleService = Provide[Container.role_service]):
     try:
+        role_service.check_superuser_authorization(user_id)
         role_list = role_service.get_roles_list()
     except ServiceException as err:
         return make_response(jsonify(err), HTTPStatus.BAD_REQUEST)
@@ -24,12 +30,17 @@ def get_roles(role_service: RoleService = Provide[Container.role_service]):
 
 
 @role.route('/', methods=['POST'])
+@jwt_required()
+@authenticate()
 @inject
-def create_role(role_service: RoleService = Provide[Container.role_service]):
+def create_role(
+        user_id: str,
+        role_service: RoleService = Provide[Container.role_service]):
     create_request = role_service.validate_role_request(request)
     if isinstance(create_request, Response):
         return create_request
     try:
+        role_service.check_superuser_authorization(user_id)
         new_role = role_service.create_role(create_request.role_name)
     except ServiceException as err:
         return make_response(jsonify(err), HTTPStatus.BAD_REQUEST)
@@ -41,13 +52,16 @@ def create_role(role_service: RoleService = Provide[Container.role_service]):
 
 
 @role.route('/<uuid:role_uuid>', methods=['PATCH'])
+@jwt_required()
+@authenticate()
 @inject
-def edit_role(role_uuid: str,
+def edit_role(user_id: str, role_uuid: str,
               role_service: RoleService = Provide[Container.role_service]):
     edit_request = role_service.validate_role_request(request)
     if isinstance(edit_request, Response):
         return edit_request
     try:
+        role_service.check_superuser_authorization(user_id)
         edited_role = role_service.edit_role(
             role_id=role_uuid,
             role_name=edit_request.role_name
@@ -62,10 +76,13 @@ def edit_role(role_uuid: str,
 
 
 @role.route('/<uuid:role_uuid>', methods=['DELETE'])
+@jwt_required()
+@authenticate()
 @inject
-def delete_role(role_uuid: str,
+def delete_role(user_id: str, role_uuid: str,
                 role_service: RoleService = Provide[Container.role_service]):
     try:
+        role_service.check_superuser_authorization(user_id)
         deleted_role = role_service.delete_role(role_id=role_uuid)
     except ServiceException as err:
         return make_response(jsonify(err), HTTPStatus.BAD_REQUEST)
@@ -76,11 +93,15 @@ def delete_role(role_uuid: str,
 
 
 @role.route('/<uuid:role_uuid>/permissions', methods=['GET'])
+@jwt_required()
+@authenticate()
 @inject
 def get_role_permissions(
+        user_id: str,
         role_uuid: str,
         role_service: RoleService = Provide[Container.role_service]):
     try:
+        role_service.check_superuser_authorization(user_id)
         perm_list = role_service.get_role_permissions(role_uuid)
     except ServiceException as err:
         return make_response(jsonify(err), HTTPStatus.BAD_REQUEST)
@@ -90,14 +111,18 @@ def get_role_permissions(
 
 
 @role.route('/<uuid:role_uuid>/permissions', methods=['POST'])
+@jwt_required()
+@authenticate()
 @inject
 def set_role_permissions(
+        user_id: str,
         role_uuid: str,
         role_service: RoleService = Provide[Container.role_service]):
     set_request = role_service.validate_perm_request(request)
     if isinstance(set_request, Response):
         return set_request
     try:
+        role_service.check_superuser_authorization(user_id)
         new_perm: Permission = role_service.set_role_permissions(
             role_id=role_uuid,
             perm_id=set_request.permission_uuid)
@@ -112,12 +137,16 @@ def set_role_permissions(
 
 @role.route('/<uuid:role_uuid>/permissions/<uuid:perm_uuid>',
             methods=['DELETE'])
+@jwt_required()
+@authenticate()
 @inject
 def remove_role_permissions(
+        user_id: str,
         role_uuid: str,
         perm_uuid: str,
         role_service: RoleService = Provide[Container.role_service]):
     try:
+        role_service.check_superuser_authorization(user_id)
         deleted_perm: Permission = role_service.remove_role_permissions(
             role_id=role_uuid,
             perm_id=perm_uuid
